+++ ChangeLog.md
@@ -17,6 +17,13 @@ Changes:
 
 * Implement `FusedIterator` for `ResultSet`.
 
+## 0.2.1 (2019-04-14)
+
+Changes:
+
+* Fix memory corruption when using object and collection data types.
+* Update ODPI-C to version 3.1.3.
+
 ## 0.2.0 (2018-10-02)
 
 Incompatible changes: